Adjuster refinery. Log scales! Breaking changes...
Adjuster can now use logarithmic scale with a customizable log base.
For now, both values get locked on creation, so you can't switch between
log and non-log freely, or change the scale. It should be possible to
have this functionality, but right now the use case is so edge I'm not
going to test all that in one batch.

Breaking changes include the creation. Adjuster() is now created using a
Gtk.Adjustment() instead of a bunch of vals. That functionality has been
moved to Adjuster.new(). A few kwargs were also renamed to not be
confusing comapred to vanilla Gtk, notably
slider -> scale, spinner -> spin_button.