First abc branch implementation
WidgetMixIn - a mix-in class to provide the common properties I use.
Might rename __init__ as it's always called manually. Or just set the
self stuff by hand, maybe, as the mix-in's main job is the properties
and methods.
DataWidgetMixIn - inherets WidgetMixIn and also adds fields for data
entry, aka most of the widgets present. the .value property of couse,
which is the first reason I made bszgw, and other newer methods.
WidgetMixInMeta - metaclass combining ABCMeta with gobject and checking
for unreplaced abstractmethods.